requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.13
django-oauth-toolkit==1.2.0
boto3
django_rest_passwordreset
packaging
django-jet
fcm-django
django-import-export==1.2.0
cloudinary==1.21.1
django-cloudinary-storage==0.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-emoticons==1.2
django-ckeditor==5.7.1
django-filebrowser-no-grappelli==3.7.8
django-filter==2.1.0
drf-yasg==1.16.1
django-rest-swagger==2.2.0 
django-haystack==2.8.1
django-markdown-deux==1.0.5
django-multiselectfield==0.1.8
django-rest-multiple-models==2.1.3
django-tagging==0.4.6
django-taggit==1.1.0
django-taggit-labels==0.6.0
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
django-richtextfield==1.4.0
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.4.0
idna==2.8
isort==4.3.20
jsmin==2.2.2
lazy-object-proxy==1.4.1
Markdown==3.1.1
markdown2==2.3.8
mccabe==0.6.1
oauthlib==3.0.1
Pillow==6.0.0
psycopg2-binary==2.8.2
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
pysolr==3.8.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.2
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.3
wrapt==1.11.1
psycopg2==2.7.2
gunicorn==19.7.1
django-froala-editor==3.0.3
pyotp
twilio